extract persistence.ts from panels/filters.ts
Move STORAGE_KEY, loadPinnedFilters, savePinnedFilters, and
isFilterShape into panels/filters/persistence.ts — the only file in
the slice that touches localStorage. panels/filters.ts re-exports the
load/save pair so callers are unchanged.
